implement item show functionality

+++ api/views/properties.py
@@ -14,7 +14,9 @@ def create():
   profile = read_token(request)
   data["profile_id"] = profile["id"]
   property = Property(**data)
+
   db.session.add(property)
+
   db.session.commit()
   return jsonify(property.serialize()), 201
 
@@ -57,15 +59,15 @@ def delete(id):
   db.session.commit()
   return jsonify(message="Success"), 200
 
-#POST api/properties/<id>/items
 @properties.route('/<id>/items', methods=["POST"])
 @login_required 
 def add_item(id):
+  property = Property.query.filter_by(id=id).first()
+  profile = read_token(request)    
+
   data = request.get_json()
   data["property_id"] = id 
-
-  profile = read_token(request)    
-  property = Property.query.filter_by(id=id).first()
+  data["profile_id"] = property.profile_id 
   
   if property.profile_id != profile["id"]:
     return 'Forbidden', 403
@@ -85,15 +87,22 @@ def item_index(property_id):
   items = Item.query.all()
   return jsonify([item.serialize() for item in items]), 200
 
+@properties.route('/<property_id>/items/<id>', methods=["GET"])
+def item_show(property_id, id):
+  item = Item.query.filter_by(id=id).first()
+  item_data = item.serialize()
+
+  return jsonify(item=item_data), 200
+
+
 @properties.route("/<property_id>/items/<id>", methods=["DELETE"])
 @login_required
 def item_delete(property_id, id):
-#  profile = read_token(request)
+  profile = read_token(request)
   item = Item.query.filter_by(id=id).first()
-#  property = Property.query.filter_by(id=id).first()
-#
-#  if property.profile_id != profile["id"]:
-#    return 'Forbidden', 403 
+
+  if item.profile_id != profile["id"]:
+    return 'Forbidden', 403 
 
   db.session.delete(item)
   db.session.commit()
